refactor: 拆分设置界面剩余的hooks和config
